requirements.txt
Django==1.11
Pillow==4.2.1
dj-database-url==0.4.2
djangorestframework==3.6.3
gunicorn==19.7.1
olefile==0.44
psycopg2==2.7.3
pytz==2017.2
